+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Note: Breaking changes between versions are indicated by "💥".
 
 ## Latest
 
-- [Improvement] Move Xqueue to a dedicated plugin
+- [Improvement] Move Xqueue and Student notes to a dedicated plugin
 
 ## 3.4.3 (2019-06-24)